Ignore newline difference in tester.py #20
Read an answer file and an output from a program in universal newline
mode. In universal newline mode, all valid newlines (i.e. \n, \r and
\r\n) are replaced by \n.
https://www.python.org/dev/peps/pep-0278/

For `open` function, universal newline mode is turned on if a stream
opened in text mode and `newline` parameter is `None`.
https://docs.python.org/3.3/library/functions.html#open

For `subprocess.check_output` function, universal newline mode is tuned
on if `universal_newlines` flag is set as `True`.
https://docs.python.org/3.6/library/subprocess.html#subprocess.check_output